Initial implementation or final if it works.
